+++ src/lib/date.spec.ts
@@ -11,18 +11,16 @@ import {
   getFirstDayOfYear,
   getLastDayOfYear,
   getQuarter,
+  getPreviousQuarter,
 } from "./date";
 
 describe("date tests", () => {
   test.each([
-    // Valid dates
     [new Date(), true],
     [new Date(new Date()), true],
     [new Date("2014-03-15"), true],
     [new Date(2014, 3, 15), true],
     [new Date(42), true],
-
-    // Invalid dates
     [new Date(Number.MAX_VALUE), false],
     [new Date(Number.NaN), false],
     [new Date("2014-03-36"), false],
@@ -176,5 +174,25 @@ describe("date tests", () => {
     [new Date("2026-11-05"), 4],
   ])("getQuarter", (date, expected) => {
     expect(getQuarter(date)).toBe(expected);
+  });
+
+  test.each([[null as unknown as Date], [undefined as unknown as Date], [42 as unknown as Date], ["test" as unknown as Date]])(
+    "getQuarter invalid inputs",
+    (date) => {
+      expect(getQuarter(date)).toBeNaN();
+    },
+  );
+
+  test.each([
+    [new Date("2026-01-15"), 4],
+    [new Date("2026-04-10"), 1],
+    [new Date("2026-08-20"), 2],
+    [new Date("2026-11-05"), 3],
+  ])("getPreviousQuarter", (date, expected) => {
+    expect(getPreviousQuarter(date)).toBe(expected);
+  });
+
+  test.each([null, undefined, 42, "test", new Date("invalid-date")] as unknown as Date[])("getPreviousQuarter invalid inputs", (date) => {
+    expect(getPreviousQuarter(date)).toBeNaN();
   });
 });